third_party/blink/web_tests/platform/generic/fast/innerHTML/innerHTML-svg-read-expected.txt

Test that accessing innerHTML and outerHTML properties works on SVG elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS innerHTML("emptysvg") is ""
PASS innerHTML("rectsvg") is "<rect></rect>"
PASS innerHTML("foreign") is "<foreignObject width=\"100\" height=\"30\"><p>html</p></foreignObject>"
PASS outerHTML("emptysvg") is "<svg id=\"emptysvg\"></svg>"
PASS outerHTML("rectsvg") is "<svg id=\"rectsvg\"><rect></rect></svg>"
PASS successfullyParsed is true

TEST COMPLETE
